Create a Guessing Game
Create a Guessing Game

Build a program that generates a random number and challenges the user to guess it. 
The program should prompt the user to input their guess, compare it to the generated number, 
and provide feedback if the guess is too high or too low. It should continue until the user 
correctly guesses the number and then display the number of attempts it took to win the game.
Filename: C:\Users\Shaibaj\eclipse-workspace\PRODIGY_SD_02\src\task_02\TemperatureConverter.java
 Path: C:\Users\Shaibaj\eclipse-workspace\PRODIGY_SD_02\src\task_02\TemperatureConverter.java
 Created Date: Monday, September 30 2024, 09:38:43 am
 Author: Shaibaj
 Copyright (c) 2024